refactor(#6): add Logo, make header sticky
