+++ app/page.tsx
@@ -2,12 +2,8 @@ import { Map } from '@/components/map/Map'
 
 export default function Home() {
   return (
-    <>
-      <header></header>
-      <main>
-        <Map />
-      </main>
-      <footer></footer>
-    </>
+    <main>
+      <Map />
+    </main>
   )
 }
